Memory optimization

+++ GGM/GGMNode.h
@@ -11,16 +11,16 @@ extern "C" {
 
 class GGMNode {
 public:
-    long index;
+    int index;
     int level;
     uint8_t key[AES_BLOCK_SIZE]{};
 
-    GGMNode(long index, int level) {
+    GGMNode(int index, int level) {
         this->index = index;
         this->level = level;
     }
 
-    GGMNode(long index, int level, uint8_t *key) {
+    GGMNode(int index, int level, uint8_t *key) {
         this->index = index;
         this->level = level;
         memcpy(this->key, key, AES_BLOCK_SIZE);
